// src/types/ErrorCode.ts
export enum ErrorCode {
  INVALID_WALLET = "INVALID_WALLET",
  INVALID_PARAMS = "INVALID_PARAMS",
  PARAM_OUT_OF_RANGE = "PARAM_OUT_OF_RANGE",
  NO_ACTIVITY = "NO_ACTIVITY",
  RPC_ERROR = "RPC_ERROR",
  PROOF_GENERATION_FAILED = "PROOF_GENERATION_FAILED",
  PROOF_VERIFICATION_FAILED = "PROOF_VERIFICATION_FAILED",
  VERIFIER_NOT_SET = "VERIFIER_NOT_SET",
  INTERNAL_ERROR = "INTERNAL_ERROR"
}